discord-copilot-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
supabase==2.9.0
discord.py==2.3.2
pydantic==2.12.5
pydantic-settings==2.1.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
openai==1.6.1  # Used for OpenRouter API (OpenAI-compatible)
pgvector==0.2.4
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
httpx==0.27.2
websockets==13.1
aiofiles==23.2.1
